requirements.txt
certifi==2020.6.20
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.0
flask-swagger-ui==3.25.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psycopg2==2.8.5
six==1.15.0
SQLAlchemy==1.3.3
validate-email==1.3
Werkzeug==1.0.1
Flask-Script==2.0.6
Flask_Migrate==2.5.3
python-jose==3.1.0
python-jose-cryptodome==1.3.2
authlib==0.14.3
requests==2.24.0
flake8==3.8.3